axing-xml-parser: Swithch from read_line to read_upto(" ")
There were a few problems with using read_line to get the chunks
of data to parse. There's no way to determine if the stream ends
with a newline, which causes conformance problems. Also, it was
difficult to track multi-char newline sequences when the chars
weren't in the same parse chunk.

This uncovered a bug in g_data_input_stream_read_upto_finish,
which I've worked around. (bugs.gnome.org #692101)