parser: Mechanical refactoring
In order to use G_DECLARE_* macro we need to remove the `priv` pointer
from the instance structure; this means that we need to retrieve the
instance private data when needed.

This is mostly a mechanical replacement of every instance of:

  context->parser->priv->some_field

With the functionally equivalent of:

  AxingXmlParserPrivate *priv =
    axing_xml_parser_get_instance_private (context->parser);
  priv->some_field

While it's a bit longer, it is more readable and it reduces the amount
of pointer dereferencing quite a bit.