Starting to do UTF-8 directly
Most of the parser involves stepping thru UTF-8 strings character by
character, calling g_utf8_get_char to get a code point to check if
it's in some ranges, then calling g_utf8_next_char. This means each
character has the overhead of two function calls, one of which has
expensive multiplication.

Instead, let's create some macros that check the UTF-8 string directly
to see if the next character is in the ranges we care about, returning
the number of bytes if it is, so we only have to compute once. And
let's use faster bit operations whenever possible.

This commit is just qnames and pubids so far. I still need to convert
the code for character data, which is where the real gains should be.