Adding an event pool
For most documents, the actual number of Event structs that need to be
in use at one time is quite low, but the total number of Event structs
that get allocated can be very large. This introduces an array of Event
structs that get reused. If we need more, we go ahead and g_new them.

The size of the pool is semi-arbitrarily set to 32. We do linearly scan
the list, but we do it somewhat smartly. Setting it too high could result
in a slowdown if the array gets fragmented, plus it obviously uses a fixed
amount of memory. Setting it too low means it has no real effect. 32 seems
like a good number based on my test documents and a few real-world XML
files. Further profiling with real-world documents could reveal a better
pool size.

This had a small but measurable speed improvement on my timing test files.