axing-xml-parser: Only use one GString, and keep reusing it
Previously we instantiated a new GString for everything we built strings for:
text content, CDATA, processing instructions, comments, and attribute values.
But we only need to accumulate one of these at a time, so we can just reuse
the same GString and truncate when done. Also, except for attribute values,
we don't need these strings to outlive the GString, so there's no need to
dup or steal the string. Just serve it from cur_text->str.

This change results in a pretty massive reduction in the number of allocs,
and it's kind of obvious in hindsight.